# openapi.yml
openapi: 3.1.0
info:
  title: Resume Copilot API
  version: 0.1.0
  description: Resume optimization and interview preparation platform
servers:
  - url: /api
    description: Local development server
paths:
...
  /tasks/{id}:
    get:
      summary: 轮询任务状态
      parameters:
        - in: path
          name: id
          required: true
          schema: 
            type: string
          description: 任务ID
      responses:
        '200':
          description: 任务状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: 任务不存在
        '500':
          description: 服务器错误
components:
  schemas:
...
    KnowledgeItem:
      type: object
      required: [title, url, why]
      properties:
        title: 
          type: string
          description: 知识点标题
        url: 
          type: string
          description: 学习链接
        why: 
          type: string
          description: 推荐理由
    Output:
      type: object
      required: [id, resume_md, interview_questions, knowledge_items]
      properties:
        id: 
          type: string
          description: 输出结果ID
        resume_md: 
          type: string
          description: 优化后的简历Markdown
        interview_questions:
          type: array
          description: 面试问题列表
          items: 
            type: string
        knowledge_items:
          type: array
          description: 知识点推荐
          items: 
            $ref: '#/components/schemas/KnowledgeItem'
    Task:
      type: object
      required: [id, status]
      properties:
        id: 
          type: string
          description: 任务ID
        status: 
          type: string
          enum: [queued, running, done, error]
          description: 任务状态
        cost: 
          type: number
          nullable: true
          description: 任务成本（元）
        latency_ms: 
          type: integer
          nullable: true
          description: 任务耗时（毫秒）
        output: 
          $ref: '#/components/schemas/Output'
          description: 任务输出结果
        error: 
          type: string
          nullable: true
          description: 错误信息